ytoolsbox-req-custom/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
grpcio==1.47.0
grpcio-tools==1.47.0
idna==3.3
protobuf==3.20.1
pyasn1==0.4.8
requests==2.28.1
rsa==4.9
six==1.16.0
supervisor==4.2.4
urllib3==1.26.10